refactor: 유저 프로필 로직 분리
